Ordo-Rating-Service-main/nlp/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.1
google-auth==2.6.4
google-cloud-core==2.3.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
idna==3.3
mysql-connector-python==8.0.28
numpy==1.22.3
pandas==1.4.2
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
